+++ examples/pullToRefresh.js
@@ -1,9 +1,10 @@
 /* eslint-disable no-console */
 /* eslint react/sort-comp: 0 */
-import 'rmc-list-view/assets/index.less';
+import '../assets/index.less';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ListView from 'rmc-list-view';
+import ListView from '../src';
+import PullToRefresh from 'rmc-pull-to-refresh';
 
 const NUM_ROWS = 6;
 
@@ -25,7 +26,8 @@ class Demo extends React.Component {
     this.state = {
       dataSource,
       useBodyScroll: true,
-      pullUpRefreshing: false,
+      refreshing: false,
+      down: false,
     };
   }
 
@@ -51,22 +53,33 @@ class Demo extends React.Component {
       >
         useBodyScroll: {this.state.useBodyScroll ? 'true' : 'false'}
       </button>
+      <button
+        style={{ display: 'inline-block', marginBottom: 30, marginLeft: 10, border: 1 }}
+        onClick={() => this.setState({ down: !this.state.down })}
+      >
+        direction: {this.state.down ? 'down' : 'up'}
+      </button>
       <ListView
         key={this.state.useBodyScroll ? 1 : 0}
         ref={el => this.lv = el}
         dataSource={this.state.dataSource}
-        renderRow={rowData => <div style={{ padding: 16 }}>{rowData}</div>}
         useBodyScroll={this.state.useBodyScroll}
         style={!this.state.useBodyScroll ? { height: 200, border: '1px solid gray' } : {}}
-        pullUpEnabled
-        pullUpRefreshing={this.state.pullUpRefreshing}
-        pullUpOnRefresh={() => {
-          this.setState({ pullUpRefreshing: true });
-          setTimeout(() => {
-            this.setState({ pullUpRefreshing: false });
-          }, 1000);
-        }}
-        pullUpRenderer={st => <div className="my-render">{st}</div>}
+        renderHeader={() => <span style={{ padding: 10 }}>header</span>}
+        renderRow={rowData => <div style={{ padding: 16 }}>{rowData}</div>}
+        pullToRefresh={
+          <PullToRefresh
+            className="forTest"
+            direction={this.state.down ? 'down' : 'up'}
+            refreshing={this.state.refreshing}
+            onRefresh={() => {
+              this.setState({ refreshing: true });
+              setTimeout(() => {
+                this.setState({ refreshing: false });
+              }, 1000);
+            }}
+          />
+        }
       />
       <div dangerouslySetInnerHTML={{
         __html: navigator.userAgent.match(/Android|iPhone|iPad|iPod/i) ?