+++ index.js
@@ -1,4 +1,0 @@
-// export this package's api
-import ListView from './src/';
-export default ListView;
-export { DataSource, IndexedList, RefreshControl } from './src';