+++ src/ScrollView.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ReactDOM from 'react-dom';
-import DOMScroller from 'zscroller';
 import classNames from 'classnames';
-import { throttle, setTransform, setTransformOrigin } from './util';
+import { throttle } from './util';
 
 // const SCROLLVIEW = 'ScrollViewRef';
 // const INNERVIEW = 'InnerScrollViewRef';
@@ -13,7 +11,6 @@ import { throttle, setTransform, setTransformOrigin } from './util';
 /* eslint react/prop-types: 0, react/sort-comp: 0, no-unused-expressions: 0 */
 
 const propTypes = {
-  children: PropTypes.any,
   className: PropTypes.string,
   prefixCls: PropTypes.string,
   listPrefixCls: PropTypes.string,
@@ -21,21 +18,6 @@ const propTypes = {
   style: PropTypes.object,
   contentContainerStyle: PropTypes.object,
   onScroll: PropTypes.func,
-  scrollEventThrottle: PropTypes.number,
-  refreshControl: PropTypes.element,
-};
-const styles = {
-  base: {
-    position: 'relative',
-    overflow: 'auto',
-    WebkitOverflowScrolling: 'touch',
-    flex: 1,
-  },
-  zScroller: {
-    position: 'relative',
-    overflow: 'hidden',
-    flex: 1,
-  },
 };
 
 export default class ScrollView extends React.Component {
@@ -46,50 +28,35 @@ export default class ScrollView extends React.Component {
     // 问题情景：用户滚动内容后，改变 dataSource 触发 ListView componentWillReceiveProps
     // 内容变化后 scrollTop 如果改变、会自动触发 scroll 事件，而此事件应该避免被执行
     if ((this.props.dataSource !== nextProps.dataSource ||
-        this.props.initialListSize !== nextProps.initialListSize) && this.tsExec) {
+        this.props.initialListSize !== nextProps.initialListSize) && this.handleScroll) {
       // console.log('componentWillUpdate');
       if (this.props.useBodyScroll) {
-        window.removeEventListener('scroll', this.tsExec);
-      } else if (!this.props.useZscroller) { // not handle useZscroller now. todo
-        ReactDOM.findDOMNode(this.ScrollViewRef).removeEventListener('scroll', this.tsExec);
+        window.removeEventListener('scroll', this.handleScroll);
+      } else {
+        this.ScrollViewRef.removeEventListener('scroll', this.handleScroll);
       }
     }
   }
   componentDidUpdate(prevProps) {
-    // console.log('componentDidUpdate');
-    if (prevProps.refreshControl && this.props.refreshControl) {
-      const preRefreshing = prevProps.refreshControl.props.refreshing;
-      const nowRefreshing = this.props.refreshControl.props.refreshing;
-      if (preRefreshing && !nowRefreshing && !this._refreshControlTimer) {
-        this.domScroller.scroller.finishPullToRefresh();
-      } else if (!this.manuallyRefresh && !preRefreshing && nowRefreshing) {
-        this.domScroller.scroller.triggerPullToRefresh();
-      }
-    }
     // handle componentWillUpdate accordingly
     if ((this.props.dataSource !== prevProps.dataSource ||
-        this.props.initialListSize !== prevProps.initialListSize) && this.tsExec) {
-      // console.log('componentDidUpdate');
+        this.props.initialListSize !== prevProps.initialListSize) && this.handleScroll) {
       setTimeout(() => {
         if (this.props.useBodyScroll) {
-          window.addEventListener('scroll', this.tsExec);
-        } else if (!this.props.useZscroller) { // not handle useZscroller now. todo
-          ReactDOM.findDOMNode(this.ScrollViewRef).addEventListener('scroll', this.tsExec);
+          window.addEventListener('scroll', this.handleScroll);
+        } else {
+          this.ScrollViewRef.addEventListener('scroll', this.handleScroll);
         }
       }, 0);
     }
-
-    // for pullUp
-    if (this.props.pullUpEnabled) {
-      const preRefreshing = prevProps.pullUpRefreshing;
-      const nowRefreshing = this.props.pullUpRefreshing;
-      if (preRefreshing && !nowRefreshing && !this._pullUpTimer) {
-        this.pullUpFinish();
-      }
-    }
   }
   componentDidMount() {
-    this.tsExec = this.throttleScroll();
+    let handleScroll = e => this.props.onScroll && this.props.onScroll(e, this.getMetrics());
+    if (this.props.scrollEventThrottle) {
+      handleScroll = throttle(handleScroll, this.props.scrollEventThrottle);
+    }
+    this.handleScroll = handleScroll;
+
     // IE supports onresize on all HTML elements.
     // In all other Browsers the onresize is only available at the window object
     this.onLayout = () => this.props.onLayout({
@@ -97,146 +64,69 @@ export default class ScrollView extends React.Component {
     });
 
     if (this.props.useBodyScroll) {
-      window.addEventListener('scroll', this.tsExec);
+      window.addEventListener('scroll', this.handleScroll);
       window.addEventListener('resize', this.onLayout);
-      this.initPullUp(document.body); // for pullUp
-    } else if (this.props.useZscroller) {
-      this.renderZscroller();
     } else {
-      const ele = ReactDOM.findDOMNode(this.ScrollViewRef);
-      ele.addEventListener('scroll', this.tsExec);
-      this.initPullUp(ele); // for pullUp
+      this.ScrollViewRef.addEventListener('scroll', this.handleScroll);
     }
   }
   componentWillUnmount() {
     if (this.props.useBodyScroll) {
-      window.removeEventListener('scroll', this.tsExec);
+      window.removeEventListener('scroll', this.handleScroll);
       window.removeEventListener('resize', this.onLayout);
-      this.destroyPullUp(document.body); // for pullUp
-    } else if (this.props.useZscroller) {
-      this.domScroller.destroy();
     } else {
-      const ele = ReactDOM.findDOMNode(this.ScrollViewRef);
-      ele.removeEventListener('scroll', this.tsExec);
-      this.destroyPullUp(ele); // for pullUp
+      this.ScrollViewRef.removeEventListener('scroll', this.handleScroll);
     }
   }
 
-  getInnerViewNode = () => {
-    return ReactDOM.findDOMNode(this.InnerScrollViewRef);
+  getMetrics = () => {
+    const isVertical = !this.props.horizontal;
+    if (this.props.useBodyScroll) {
+      // In chrome61 `document.body.scrollTop` is invalid,
+      // and add new `document.scrollingElement`(chrome61, iOS support).
+      // In old-android-browser and iOS `document.documentElement.scrollTop` is invalid.
+      const scrollNode = document.scrollingElement ? document.scrollingElement : document.body;
+      return {
+        visibleLength: window[isVertical ? 'innerHeight' : 'innerWidth'],
+        contentLength: this.ScrollViewRef[isVertical ? 'scrollHeight' : 'scrollWidth'],
+        offset: scrollNode[isVertical ? 'scrollTop' : 'scrollLeft'],
+      };
+    }
+    return {
+      visibleLength: this.ScrollViewRef[isVertical ? 'offsetHeight' : 'offsetWidth'],
+      contentLength: this.ScrollViewRef[isVertical ? 'scrollHeight' : 'scrollWidth'],
+      offset: this.ScrollViewRef[isVertical ? 'scrollTop' : 'scrollLeft'],
+    };
   }
+
+  getInnerViewNode = () => this.InnerScrollViewRef;
 
   scrollTo = (...args) => {
     if (this.props.useBodyScroll) {
       window.scrollTo(...args);
-    } else if (this.props.useZscroller) {
-      // it will change zScroller's dimensions on data loaded, so it needs fire reflow.
-      this.domScroller.reflow();
-      this.domScroller.scroller.scrollTo(...args);
     } else {
-      const ele = ReactDOM.findDOMNode(this.ScrollViewRef);
-      ele.scrollLeft = args[0];
-      ele.scrollTop = args[1];
+      this.ScrollViewRef.scrollLeft = args[0];
+      this.ScrollViewRef.scrollTop = args[1];
     }
   }
 
-  throttleScroll = () => {
-    let handleScroll = () => {};
-    if (this.props.scrollEventThrottle && this.props.onScroll) {
-      handleScroll = throttle(e => {
-        this.props.onScroll && this.props.onScroll(e);
-      }, this.props.scrollEventThrottle);
-    }
-    return handleScroll;
-  }
-
-  scrollingComplete = () => {
-    // console.log('scrolling complete');
-    if (this.props.refreshControl &&
-    this.RefreshControlRef && this.RefreshControlRef.state.deactive) {
-      this.RefreshControlRef.setState({ deactive: false });
-    }
-  }
-
-  renderZscroller() {
-    const { scrollerOptions, refreshControl } = this.props;
-    const { scrollingComplete, onScroll, ...restProps } = scrollerOptions;
-    // console.log(scrollingComplete, onScroll, restProps);
-    // console.log('onRefresh will not change', refreshControl.props.onRefresh.toString());
-    this.domScroller = new DOMScroller(this.getInnerViewNode(), {
-      scrollingX: false,
-      onScroll: () => {
-        this.tsExec();
-        if (onScroll) {
-          onScroll();
-        }
-      },
-      scrollingComplete: () => {
-        this.scrollingComplete();
-        if (scrollingComplete) {
-          scrollingComplete();
-        }
-      },
-      ...restProps,
-    });
-    if (refreshControl) {
-      const scroller = this.domScroller.scroller;
-      const { distanceToRefresh, onRefresh } = refreshControl.props;
-      scroller.activatePullToRefresh(distanceToRefresh,
-        () => {
-          // console.log('reach to the distance');
-          this.manuallyRefresh = true;
-          this.overDistanceThenRelease = false;
-          this.RefreshControlRef && this.RefreshControlRef.setState({ active: true });
-        },
-        () => {
-          // console.log('back to the distance');
-          this.manuallyRefresh = false;
-          this.RefreshControlRef && this.RefreshControlRef.setState({
-            deactive: this.overDistanceThenRelease,
-            active: false,
-            loadingState: false,
-          });
-        },
-        () => {
-          // console.log('Over distance and release to loading');
-          this.overDistanceThenRelease = true;
-          this.RefreshControlRef && this.RefreshControlRef.setState({
-            deactive: false,
-            loadingState: true,
-          });
-          this._refreshControlTimer = setTimeout(() => {
-            if (!this.props.refreshControl.props.refreshing) {
-              scroller.finishPullToRefresh();
-            }
-            this._refreshControlTimer = undefined;
-          }, 1000);
-          onRefresh();
-        });
-      if (refreshControl.props.refreshing) {
-        scroller.triggerPullToRefresh();
-      }
-    }
-  }
   render() {
     const {
       children, className, prefixCls, listPrefixCls, listViewPrefixCls,
-      style = {}, contentContainerStyle = {},
-      useZscroller, refreshControl, useBodyScroll, pullUpEnabled, pullUpRenderer,
+      style = {}, contentContainerStyle = {}, useBodyScroll, pullToRefresh,
     } = this.props;
 
-    let styleBase = styles.base;
-    if (useBodyScroll) {
-      styleBase = {};
-    } else if (useZscroller) {
-      styleBase = styles.zScroller;
-    }
-
+    const styleBase = {
+      position: 'relative',
+      overflow: 'auto',
+      WebkitOverflowScrolling: 'touch',
+      flex: 1,
+    };
     const preCls = prefixCls || listViewPrefixCls || '';
 
     const containerProps = {
       ref: el => this.ScrollViewRef = el,
-      style: { ...styleBase, ...style },
+      style: { ...(useBodyScroll ? {} : styleBase), ...style },
       className: classNames(className, `${preCls}-scrollview`),
     };
     const contentContainerProps = {
@@ -245,45 +135,16 @@ export default class ScrollView extends React.Component {
       className: classNames(`${preCls}-scrollview-content`, listPrefixCls),
     };
 
-    if (refreshControl) {
-      return (
-        <div {...containerProps}>
-          <div {...contentContainerProps}>
-            {React.cloneElement(refreshControl, { ref: el => this.RefreshControlRef = el })}
-            {children}
-          </div>
-        </div>
-      );
-    }
-
-    const createPullUp = () => {
-      const pullUpCls = classNames(`${preCls}-pull-up-content`,
-        !this.state.isTouching && `${preCls}-pull-up-dropped`);
-      let defaultRenderer = this.pullUpDisplay.deactivate;
-      switch (this.state.pullUp) {
-        case 'activate':
-        case 'deactivate':
-        case 'release':
-        case 'finish':
-        default:
-          defaultRenderer = this.pullUpDisplay[this.state.pullUp];
-      }
-      return (
-        <div className={pullUpCls} ref={el => this.pullUpContentRef = el}>
-          {children}
-          <div ref={el => this.pullUpIndicatorRef = el} className={`${preCls}-pull-up-indicator`}>
-            {pullUpRenderer ? pullUpRenderer(this.state.pullUp) : defaultRenderer}
-          </div>
-        </div>
-      );
-    };
+    const clonePullToRefresh = isBody => React.cloneElement(pullToRefresh, {
+      prefixCls: `${preCls}-pull-to-refresh`,
+      getScrollContainer: isBody ? () => document.body : () => this.ScrollViewRef,
+    }, children);
 
     if (useBodyScroll) {
-      if (pullUpEnabled) {
-        containerProps.style.overflow = 'hidden';
+      if (pullToRefresh) {
         return (
           <div {...containerProps}>
-            {createPullUp()}
+            {clonePullToRefresh(true)}
           </div>
         );
       }
@@ -294,12 +155,11 @@ export default class ScrollView extends React.Component {
       );
     }
 
-    if (pullUpEnabled) {
-      contentContainerProps.style.overflow = 'hidden';
+    if (pullToRefresh && this.ScrollViewRef) {
       return (
         <div {...containerProps}>
           <div {...contentContainerProps}>
-            {createPullUp()}
+            {clonePullToRefresh()}
           </div>
         </div>
       );
@@ -312,135 +172,5 @@ export default class ScrollView extends React.Component {
         </div>
       </div>
     );
-  }
-
-  /**
-   The following code was intended to implement the pull-up-to-load-more feature,
-
-   Coincidentally, it solves a problem, if the content is not high enough,
-   the `onScroll` and `onEndReached` event will not be fired.
-   However, there should be a better solution for this issue.
-   */
-  // https://github.com/yiminghe/zscroller/blob/2d97973287135745818a0537712235a39a6a62a1/src/Scroller.js#L355
-  // states: `activate` / `deactivate` / `release` / `finish`
-  state = {
-    pullUp: false,
-    isTouching: false,
-  };
-
-  pullUpStats = {
-    activate: 'activate',
-    deactivate: 'deactivate',
-    release: 'release',
-    finish: 'finish',
-  };
-
-  pullUpDisplay = {
-    activate: '释放刷新',
-    deactivate: '上拉 ↑',
-    release: '加载中...',
-    finish: '完成刷新',
-  };
-
-  genEvtHandler = (ele) => {
-    return {
-      touchstart: this.onTouchStart.bind(this, ele),
-      touchmove: this.onTouchMove.bind(this, ele),
-      touchend: this.onTouchEnd.bind(this, ele),
-      touchcancel: this.onTouchEnd.bind(this, ele),
-    };
-  }
-
-  initPullUp = (ele) => {
-    if (this.pullUpContentRef) {
-      setTransformOrigin(this.pullUpContentRef.style, 'left top');
-    }
-
-    this._to = this.genEvtHandler(ele);
-    Object.keys(this._to).forEach(key => {
-      ele.addEventListener(key, this._to[key]);
-    });
-  }
-
-  destroyPullUp = (ele) => {
-    Object.keys(this._to).forEach(key => {
-      ele.removeEventListener(key, this._to[key]);
-    });
-  }
-
-  onTouchStart = (ele, e) => {
-    this._pullUpScreenY = this._pullUpStartScreenY = e.touches[0].screenY;
-    this._pullUpLastScreenY = 0;
-    if (this.props.pullUpEnabled) {
-      this.setState({ isTouching: true });
-    }
-  }
-
-  onTouchMove = (ele, e) => {
-    if (!this.props.pullUpEnabled) {
-      return;
-    }
-
-    // 使用 pageY 对比有问题
-    const _screenY = e.touches[0].screenY;
-    if (this._pullUpStartScreenY - _screenY > 0) {
-      // console.log('is pull up', _screenY);
-
-      let isReachBottom;
-      if (this.props.useBodyScroll) {
-        // In chrome61 `document.body.scrollTop` is invalid, here `ele === document.body`
-        const scrollNode = document.scrollingElement ? document.scrollingElement : ele;
-        isReachBottom = ele.scrollHeight - scrollNode.scrollTop <= window.innerHeight;
-        // console.log(ele.scrollHeight, scrollNode.scrollTop, window.innerHeight);
-      } else {
-        isReachBottom = ele.scrollHeight - ele.scrollTop === ele.clientHeight;
-      }
-      if (isReachBottom) {
-        const _diff = Math.round(_screenY - this._pullUpScreenY);
-        this._pullUpScreenY = _screenY;
-        this._pullUpLastScreenY += _diff;
-
-        setTransform(this.pullUpContentRef.style,
-          `translate3d(0px,${this._pullUpLastScreenY}px,0)`);
-
-        if (Math.abs(this._pullUpLastScreenY) < this.props.pullUpDistanceToRefresh) {
-          if (this.state.pullUp !== this.pullUpStats.deactivate) {
-            // console.log('back to the distance');
-            this.setState({ pullUp: this.pullUpStats.deactivate });
-          }
-        } else {
-          if (this.state.pullUp === this.pullUpStats.deactivate) {
-            // console.log('reach to the distance');
-            this.setState({ pullUp: this.pullUpStats.activate });
-          }
-        }
-      }
-    }
-  }
-
-  onTouchEnd = () => {
-    if (this.props.pullUpEnabled) {
-      this.setState({ isTouching: false });
-    }
-    if (this.state.pullUp === this.pullUpStats.deactivate) {
-      this.pullUpFinish();
-    } else if (this.state.pullUp === this.pullUpStats.activate) {
-      this.setState({ pullUp: this.pullUpStats.release });
-      this._pullUpTimer = setTimeout(() => {
-        if (!this.props.pullUpRefreshing) {
-          this.pullUpFinish();
-        }
-        this._pullUpTimer = undefined;
-      }, 1000);
-      this.props.pullUpOnRefresh();
-    }
-  }
-
-  pullUpFinish = () => {
-    this._pullUpLastScreenY = 0;
-    setTransform(this.pullUpContentRef.style, `translate3d(0px,0px,0)`);
-    if (this.state.pullUp === this.pullUpStats.release) {
-      this.setState({ pullUp: this.pullUpStats.finish });
-    }
   }
 }