+++ src/index.js
@@ -1,10 +1,9 @@
 // export this package's api
 import ListView from './ListView';
 import IndexedList from './Indexed';
-import RefreshControl from './RefreshControl';
-ListView.IndexedList = IndexedList;
-ListView.RefreshControl = RefreshControl;
 
+ListView.IndexedList = IndexedList;
 const DataSource = ListView.DataSource;
-export { DataSource, IndexedList, RefreshControl };
+
+export { DataSource, IndexedList };
 export default ListView;